Adding pre-commit hook yaml file
This will make sure that all the code is up to standard with PEP8.
PEP8 is the "standard" for Python code, and even if its kinda weird
sometimes at least its consistent.

For longer lines, there seems to be an issue with boolean operators and
arithmetic. It doesnt like when there is a newline immediately after an
operator, nor if there is a newline immediately before an operator.

Style guides tell us that immediately before is better, because you can
see the operator and operand used right there. I agree, so I've
overriden the W503 errror manually.