Start communicating with the Google maps API
While this is is not in the greatest state, we are actually sending
requests and expecting either a failure or a response based on the usage
of a correct key.

We will likely want to stub out this client connection once we actually
get an idea of what the data looks like and how we can properly
communicate with this API. For now, the API calls are cheap, so who
cares.