Refactor the webpages and visitor
In order to allow the webpages to be parsed for different types of data,
and to allow different types data sources other than HTML webpages, we
have decided to follow the "Visitor" pattern, which inverts the
responsibilities of the previously defined objects.

Now, a Webpage can accept visitors, and the "Visitor" duck type can
visit a webpage.